feature: write middleware to check whether a user has permission to view user routes
